+++ tests/testthat/test-roxytest-tests-cls_fact_fam.R
@@ -2,13 +2,11 @@
 
 # File R/cls_fact_fam.R: @tests
 
-test_that("Function FamilyEvent() @ L57", {
+test_that("Function FamilyEvent() @ L55", {
   expect_error(FamilyEvent("marr", fact_val = "Y"), 
                regexp = "This is not a valid @fact_type for this event")
   expect_error(FamilyEvent("MARR", fact_val = "Yes"), 
                regexp = "Only a @fact_val of 'Y' is permitted for this event")
-  expect_error(FamilyEvent("EVEN", fact_desc = "Fact desc"), 
-               regexp = "A @fact_val is required for this fact")       
   expect_error(FamilyEvent("DIV", fact_val = "Y", husb_age = "73"), regexp = "@husb_age is in an invalid format")
   expect_snapshot_value(FamilyEvent("DIV", fact_val = "Y")@GEDCOM, "json2")
   expect_snapshot_value(FamilyEvent("DIV", fact_val = "Y", wife_age_phrase = "old")@GEDCOM, "json2")
@@ -17,7 +15,7 @@ test_that("Function FamilyEvent() @ L57", {
 })
 
 
-test_that("Function FamilyAttribute() @ L80", {
+test_that("Function FamilyAttribute() @ L78", {
   expect_error(FamilyAttribute("residence", fact_val = "Earth"), 
                regexp = "This is not a valid @fact_type for this attribute")
   expect_error(FamilyAttribute("RESI", fact_val = ""),