+++ NEWS.md
@@ -1,5 +1,6 @@
 # gedcomS7 development version
 
+* Bespoke events (`FamilyEvent("EVEN")` and `IndividualEvent("EVEN")`) no longer require a `fact_val` argument
 * `add_children()` and `add_siblings` gain a `pedigrees` argument allowing Child to Family links to be defined more specifically
 * `Address()` objects no longer treat semi-colons as new lines